Add Nullable struct to control marshaling of nil values
Previously, there was no way to selectively omit fields from a request when their
value is nil. Optional fields use the omitempty tag, so there was no way to
force the SDK to set the field to "null" in the JSON request.

This change adds the Nullable wrapper. The inner value is always marshaled,
even if it's nil. The Nullable itself, though, still uses omitempty, so it's
omitted if nil. It's set based on the `isNullable` property, which is an
OpenAPI 3.0 keyword. I'm going to introduce a followup change to set it
based on an x-nullable tag when upgrading our 2.0 spec to 3.0.

There was previously a set of Nullable classes for primitive types, but
it didn't quite do what we wanted it to, so I removed it in favor of a
generic solution.

I added an Omit* method to Nullable, but I'm not sure it's necessary. All
of the CCAPI's PATCH request objects are separate from their corresponding
response objects, so it's unlikely that anyone's ever going to need to
call Omit*. They can just leave the field out of the struct.