Use default MarshalJSON behavior
The OpenAPI generator generates a custom MarshalJSON method for each model.
The generated code adds each model field into a map, and then marshals the
map using the default JSON marshalling code. However, this generated code
doesn't respect JSON field attributes like "string", which are meant to
modify marshalling behavior. This issue is preventing the SDK from round-
tripping certain models: the UnmarshalJSON method is not overridden and
does respect JSON field attributes, while the MarshalJSON method does not
respect them. This is causing errors in the ccloud tool, which expects all
fields to be round-trippable.

The fix for this problem is to stop generating custom MarshalJSON methods,
so that the default JSON marshalling will run instead. It doesn't seem like
the custom marshalling code is doing anything special (at least not for the
models we use), so it should be safe to skip it.