Add the description field to the generation enum docs
With the addition of the Limited Access features to the latest spec,
it's important to denote these features in the documentation because you
cannot access them without requesting access.  In some cases, they are
marked as such by the description field in the open api spec. Therefore
I've updated the enum template to include this in a block quote.  The
formatting isn't the greatest currently.  Specifically, the openapi
generation process is not rendering newlines within the description.